sched/fair: Reduce no-hz idle balance for energy aware systems
We are kicking no-hz idle balance when the CPU has two tasks.
It should be done only when the CPU is overutilized on
energy aware systems. This approach seems to be causing more
number of wakeups on HZ=250 systems compared to HZ=100.

Revisit the no-hz idle scheme and only send the IPI if and
when it is necessary.

Change-Id: I82e58a2aa3b6508c86d05c93c9fb7fe6dfbde500
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
Signed-off-by: dracarys18 <karthihegde010@gmail.com>
Signed-off-by: AtomicXZ <achubadyal4@gmail.com>