kernel: Add API to affine IRQs and kthreads on certain Big CPUs
On recent devices with heterogeneous cores, it is beneficial to
have certain IRQs and kthreads affined to certain Big CPUs so that
these performance critical IRQs/kthreads doesn't janks up a single CPU
at a same time.

Heterogeneous Big.little cluster with a 6.2 (Big-2, Little-6) seems to
be most beneficial to this implementation and this API is actually
written based on it.

The commit adds the following APIs:
- IRQF_PERF_FIRST_AFFINE affines IRQs on cpu6
- IRQF_PERF_SECOND_AFFINE affines IRQs on cpu7

Signed-off-by: Panchajanya1999 <panchajanya@azure-dev.live>
Signed-off-by: AtomicXZ <achubadyal4@gmail.com>